ref: moved MobileSidebar component into /mobile dir
